Package structure and foldering redesign
* separate files for different parts of the package is created.
* Same for usages files in examples folder
