+++ examples/usage_Entering_Commands.jl
@@ -1,0 +1,20 @@
+developing=true; # are you developing the package or just using?
+if developing
+	# ] add Revise # if you don't have it
+	# ] add Requires
+	path=pwd()
+	if path[end-6:end] != "MatLang"
+		error("cd(to the MatLang path)")
+	end
+	push!(LOAD_PATH, ".")
+	using Revise
+	using MatLang
+else
+	using MatLang
+end
+################################################################
+################################################################
+
+################################################################
+clcM() # clears console
+################################################################
