+++ test/runtests.jl
@@ -1,12 +1,21 @@
-# path=pwd()
-# if path[end-6:end] != "MatLang"
-# 	error("cd(to the MatLang path)")
-# end
+developing=true; # are you developing the package or just using?
+if developing
+	# ] add Revise # if you don't have it
+	# ] add Requires
+	path=pwd()
+	if path[end-6:end] != "MatLang"
+		error("cd(to the MatLang path)")
+	end
+	push!(LOAD_PATH, ".")
+	using Revise
+	using MatLang
+	using Test
 
-push!(LOAD_PATH, ".")
-using Revise
-using MatLang
-using Test
+else
+	using MatLang
+	using Test
+end
+################################################################
 
 # tests
 @testset "MatLang.jl" begin
